unnecessary code cleaning...

+++ psana/psana/graphqt/lcls1/GUViewHist.py
@@ -10,7 +10,7 @@ Usage ::
     import sys
     from PyQt4 import QtGui, QtCore
     from graphqt.GUViewHist import GUViewHist
-    rectax=QtCore.QRectF(0, -1.2, 100, 2.4)    
+    rectax=QtCore.QRectF(0, -1.2, 100, 2.4)
     app = QtGui.QApplication(sys.argv)
     w = GUViewHist(None, rectax, origin='DL', scale_ctl='H', rulers='DL',\
                     margl=0.12, margr=0.02, margt=0.02, margb=0.06)
@@ -25,7 +25,7 @@ Usage ::
     w.connect_statistics_updated_to(recip)
     w.disconnect_statistics_updated_from(recip)
     w.test_statistics_std_reception(self, mean, rms, err_mean, err_rms, neff, skew, kurt, err_err, sum_w)
- 
+
     w.connect_mean_std_updated_to(recip)
     w.disconnect_mean_std_updated_from(recip)
     w.test_hist_mean_std_updated_reception(self, mean, rms)
@@ -52,7 +52,7 @@ Usage ::
 
     ind, hb.values = w.hist_bin_value(x, ihis=0)
 
-    w.reset_original_hist() 
+    w.reset_original_hist()
     w.add_hist(values, edges, pen=QtGui.QPen(Qt.black), brush=QtGui.QBrush(), vtype=np.float)
     w.add_array_as_hist(arr, pen=QtGui.QPen(Qt.black), brush=QtGui.QBrush(), vtype=np.float)
     w.remove_all_graphs()
@@ -67,7 +67,7 @@ Usage ::
 
     Re-defines methods
     ------------------
-    mouseReleaseEvent, closeEvent, mouseMoveEvent, wheelEvent, 
+    mouseReleaseEvent, closeEvent, mouseMoveEvent, wheelEvent,
     enterEvent, leaveEvent, keyPressEvent, __del__
 
     Global scope methods
@@ -86,10 +86,9 @@ from pyimgalgos.HBins import HBins
 import numpy as np
 import math
 
-#------------------------------
 
 def image_to_hist_arr(arr, vmin=None, vmax=None, nbins=None) :
-    
+
     amin = math.floor(arr.min() if vmin is None else vmin)
     amax = math.ceil (arr.max() if vmax is None else vmax)
 
@@ -102,11 +101,11 @@ def image_to_hist_arr(arr, vmin=None, vmax=None, nbins=None) :
     if amin == amax : amax += 1
     nhbins = int(amax-amin) if nbins is None else nbins
 
-    NBINS_MIN = 2 if arr.dtype == np.int else 100
+    NBINS_MIN = 2 if arr.dtype == np.int32 else 100
     NBINS_MAX = (1<<15) - 1
 
     #print 'XXX:NBINS_MAX', NBINS_MAX
-    
+
     if nhbins>NBINS_MAX : nhbins=NBINS_MAX
     if nhbins<NBINS_MIN : nhbins=NBINS_MIN
 
@@ -119,15 +118,14 @@ def image_to_hist_arr(arr, vmin=None, vmax=None, nbins=None) :
 
     return amin, amax, nhbins, values
 
-#------------------------------
 
 class GUViewHist(GUViewAxes) :
-    
+
     def __init__(self, parent=None, rectax=QtCore.QRectF(0, 0, 1, 1), origin='DL', scale_ctl='HV', rulers='TBLR',\
                  margl=None, margr=None, margt=None, margb=None) :
 
-        #xmin, xmax = np.amin(x), np.amax(x) 
-        #ymin, ymax = np.amin(y), np.amax(y) 
+        #xmin, xmax = np.amin(x), np.amax(x)
+        #ymin, ymax = np.amin(y), np.amax(y)
         #w, h = xmax-xmin, ymax-ymin
 
         GUViewAxes.__init__(self, parent, rectax, origin, scale_ctl, rulers, margl, margr, margt, margb)
@@ -136,7 +134,7 @@ class GUViewHist(GUViewAxes) :
 
         self.countemit = 0
         self.ibin_old = None
-        
+
         self.do_cursor_scope = True
         if self.do_cursor_scope : self._init_cursor_scope()
 
@@ -178,7 +176,6 @@ class GUViewHist(GUViewAxes) :
         #self.setWindowFlags(self.windowFlags() | QtCore.Qt.FramelessWindowHint)
         #self.setAttribute(Qt.WA_TranslucentBackground)
 
-#------------------------------
 
     def set_limits_horizontal(self, amin=None, amax=None) :
         self.amin, self.amax = amin, amax
@@ -201,13 +198,13 @@ class GUViewHist(GUViewAxes) :
 
         for hb in self.lst_hbins :
             i1,i2 = hb.bin_indexes((x1,x2))
-            hmin = hb.values[i1] if i1==i2 else hb.values[i1:i2].min()            
+            hmin = hb.values[i1] if i1==i2 else hb.values[i1:i2].min()
             hmax = hb.values[i1] if i1==i2 else hb.values[i1:i2].max()
             hmean= hb.values[i1] if i1==i2 else hb.values[i1:i2].mean()
             if hmax>20*hmean : hmax = 20*hmean
 
-            ymin = hmin #if ymin is None else min(hmin,ymin) 
-            ymax = hmax + 0.12*(hmax-hmin) #if ymax is None else max(hmax,ymax) 
+            ymin = hmin #if ymin is None else min(hmin,ymin)
+            ymax = hmax + 0.12*(hmax-hmin) #if ymax is None else max(hmax,ymax)
             #print hb.values
 
         #print 'i1, i2, hmin, hmax:', i1, i2, hmin, hmax
@@ -216,7 +213,6 @@ class GUViewHist(GUViewAxes) :
 
         self.evaluate_hist_mean_std()
 
-#------------------------------
 
     def visible_hist_vce(self, ihis=0):
         '''Returns arrays of values, bin centers, and edges for visible part of the hystogram
@@ -239,10 +235,10 @@ class GUViewHist(GUViewAxes) :
         #print_ndarr(hb.values, name='XXX: hb.values()', first=0, last=10)
         #print_ndarr(hb.bincenters(), name='XXX: hb.bincenters()', first=0, last=10)
         #print_ndarr(hb.binedges(), name='XXX: hb.binedges()', first=0, last=10)
-       
+
         values  = hb.values[i1:i2]
         centers = hb.bincenters()[i1:i2]
-        edges   = hb.binedges()[i1:i2+1] 
+        edges   = hb.binedges()[i1:i2+1]
 
         #print 'XXX: GUViewHist.visible_hist_vce'
         #print 'XXX: hb.nbins()', hb.nbins()
@@ -278,14 +274,12 @@ class GUViewHist(GUViewAxes) :
         #print 'XXX visible_hist_mean_std:', mean, std
         return mean, std
 
-#------------------------------
 
     def on_axes_limits_changed(self, x1, x2, y1, y2):
-        #print 'XXX:GUViewHist.on_axes_limits_changed  x1: %.2f  x2: %.2f  y1: %.2f  y2: %.2f' % (x1, x2, y1, y2)      
+        #print 'XXX:GUViewHist.on_axes_limits_changed  x1: %.2f  x2: %.2f  y1: %.2f  y2: %.2f' % (x1, x2, y1, y2)
         self.set_limits_horizontal(amin=x1, amax=x2)
         self.evaluate_hist_statistics()
 
-#------------------------------
 
     def evaluate_hist_statistics(self):
         """Evaluates histogram statistical parameters and emits them with signal"""
@@ -305,7 +299,6 @@ class GUViewHist(GUViewAxes) :
               'mean, rms, err_mean, err_rms, neff, skew, kurt, err_err, sum_w\n',\
                mean, rms, err_mean, err_rms, neff, skew, kurt, err_err, sum_w
 
-#------------------------------
 
     def evaluate_hist_mean_std(self):
         """Evaluates histogram mean and std (standard deviation) and emits them with signal"""
@@ -323,7 +316,6 @@ class GUViewHist(GUViewAxes) :
     def test_hist_mean_std_updated_reception(self, mean, rms) :
         print 'GUViewHist.test_hist_mean_std_updated_reception mean: %.2f  rms: %.2f' % (mean, rms)
 
-#------------------------------
 
     def hist_bin_value(self, x, ihis=0):
         '''Returns arrays of values, bin centers, and edges for visible part of the hystogram
@@ -333,14 +325,12 @@ class GUViewHist(GUViewAxes) :
         ind = hb.bin_index(x)
         return ind, hb.values[ind]
 
-#------------------------------
- 
+
     def mouseReleaseEvent(self, e):
         GUViewAxes.mouseReleaseEvent(self, e)
         #print 'GUViewHist.mouseReleaseEvent'
 
-#------------------------------
- 
+
     def closeEvent(self, e):
         #print 'GUViewHist.closeEvent'
         self.lst_items = []
@@ -348,25 +338,24 @@ class GUViewHist(GUViewAxes) :
         GUViewAxes.closeEvent(self, e)
         #print '%s.closeEvent' % self._name
 
-#------------------------------
 
     def mouseMoveEvent(self, e):
         GUViewAxes.mouseMoveEvent(self, e) # calls display_pixel_pos(e)
 
-        if self.do_cursor_scope : 
+        if self.do_cursor_scope :
             p = self.mapToScene(e.pos())
             x, y = p.x(), p.y()
             if x<self.rectax.left() : return
             y1, y2 = self.rectax.bottom(), self.rectax.top()
             self.cline1i.setLine(x, y1, x, y2)
             self.cline2i.setLine(x, y1, x, y2)
-            
+
             ibin, val = self.hist_bin_value(x)
             if ibin != self.ibin_old :
                 #print 'x, ibin, val', x, ibin, val
                 self.ibin_old = ibin
                 self.emit(QtCore.SIGNAL('cursor_bin_changed(float,float,float,float)'), x, y, ibin, val)
-               
+
         if self.pos_click is None : return
         self.set_limits()
 
@@ -376,7 +365,6 @@ class GUViewHist(GUViewAxes) :
     def disconnect_cursor_bin_changed_from(self, recip) :
         self.disconnect(self, QtCore.SIGNAL('cursor_bin_changed(float,float,float,float)'), recip)
 
-#------------------------------
 
     def wheelEvent(self, e) :
         GUViewAxes.wheelEvent(self, e)
@@ -386,15 +374,15 @@ class GUViewHist(GUViewAxes) :
     def enterEvent(self, e) :
     #    print 'enterEvent'
         GUViewAxes.enterEvent(self, e)
-        if self.do_cursor_scope : 
+        if self.do_cursor_scope :
             self.cline1i.setPen(self.pen1)
             self.cline2i.setPen(self.pen2)
-        
+
 
     def leaveEvent(self, e) :
     #    print 'leaveEvent'
         GUViewAxes.leaveEvent(self, e)
-        if self.do_cursor_scope : 
+        if self.do_cursor_scope :
             self.cline1i.setPen(QtGui.QPen())
             self.cline2i.setPen(QtGui.QPen())
 
@@ -406,11 +394,11 @@ class GUViewHist(GUViewAxes) :
                '\n'
 
     def keyPressEvent(self, e) :
-        #print 'keyPressEvent, key=', e.key()         
+        #print 'keyPressEvent, key=', e.key()
         if   e.key() == Qt.Key_Escape :
             self.close()
 
-        elif e.key() == Qt.Key_R : 
+        elif e.key() == Qt.Key_R :
             self.reset_original_hist()
 
         elif e.key() == Qt.Key_N :
@@ -418,15 +406,15 @@ class GUViewHist(GUViewAxes) :
             print '%s: Test set new histogram' % self._name
             arr = image_with_random_peaks((50, 50))
             self.remove_all_graphs()
-            hcolor = Qt.green # Qt.yellow Qt.blue Qt.yellow 
+            hcolor = Qt.green # Qt.yellow Qt.blue Qt.yellow
             self.add_array_as_hist(arr, pen=QtGui.QPen(hcolor, 0), brush=QtGui.QBrush(hcolor))
         else :
             print self.key_usage()
 
-    
+
     def reset_original_hist(self) :
 
-        #print 'GUViewHist.reset_original_hist'        
+        #print 'GUViewHist.reset_original_hist'
 
         #print 'Reset original size'
         #ihis=0
@@ -447,7 +435,6 @@ class GUViewHist(GUViewAxes) :
         self.lst_items.append(self.scene().addPath(path, pen, brush))
         #self.update_my_scene() # ?????
 
-#------------------------------
 
     def add_hist(self, values, edges, pen=QtGui.QPen(Qt.black), brush=QtGui.QBrush(), vtype=np.float) :
         nbins = len(values) #if isinstance(values, (list,tuple)) else values.size
@@ -478,7 +465,6 @@ class GUViewHist(GUViewAxes) :
 
         self.emit(QtCore.SIGNAL('histogram_updated()'))
 
-#------------------------------
 
     def connect_histogram_updated_to(self, recip) :
         self.connect(self, QtCore.SIGNAL('histogram_updated()'), recip)
@@ -489,14 +475,13 @@ class GUViewHist(GUViewAxes) :
     def test_histogram_updated_reception(self) :
         print 'GUViewHist.test_histogram_updated_reception'
 
-#------------------------------
 
     def add_array_as_hist(self, arr, pen=QtGui.QPen(Qt.black), brush=QtGui.QBrush(),\
                           vtype=np.float, set_hlims=True) :
         """Adds array (i.e. like image) as a histogram of intensities
         """
         amin, amax, nbins, values = image_to_hist_arr(arr) #, self.amin, self.amax)
-        if set_hlims : 
+        if set_hlims :
             self.set_limits_horizontal(amin, amax)
 
         vmin, vmax = values.min(), values.max()
@@ -506,7 +491,7 @@ class GUViewHist(GUViewAxes) :
         #print 'XXX: GUViewHist.add_array_as_hist: amin=%.1f  amax=%.1f  vmin=%.1f  vmax=%.1f' % (amin, amax, vmin, vmax)
 
         self.set_limits_vertical(None, vmax)
-        self.raxes = QtCore.QRectF(self.amin, vmin, self.amax-self.amin, vmax-vmin) 
+        self.raxes = QtCore.QRectF(self.amin, vmin, self.amax-self.amin, vmax-vmin)
 
         #print 'XXX A'
         self.add_hist(values, (amin,amax), pen, brush, vtype)
@@ -525,7 +510,7 @@ class GUViewHist(GUViewAxes) :
     def __del__(self) :
         self.remove_all_graphs()
 
-#------------------------------
+
 
 if __name__ == "__main__" :
 
@@ -540,14 +525,14 @@ if __name__ == "__main__" :
     y2 = np.sin(x2)/x2
     #y2 = np.random.random((nbins,))
 
-    rectax=QtCore.QRectF(0, -1.2, 100, 2.4)    
+    rectax=QtCore.QRectF(0, -1.2, 100, 2.4)
     app = QtGui.QApplication(sys.argv)
     w = GUViewHist(None, rectax, origin='DL', scale_ctl='H', rulers='DL',\
                     margl=0.12, margr=0.02, margt=0.02, margb=0.06)
 
-    w.set_limits_vertical(-1.2, 1.2) 
+    w.set_limits_vertical(-1.2, 1.2)
     w.setWindowTitle("GUViewHist")
-    
+
     w.connect_axes_limits_changed_to(w.test_axes_limits_changed_reception)
     #w.disconnect_axes_limits_changed_from(w.test_axes_limits_changed_reception)
 
@@ -566,4 +551,4 @@ if __name__ == "__main__" :
     w.show()
     app.exec_()
 
-#-----------------------------
+# EOF
